Tested endpoints, finished backend

+++ donations/models.py
@@ -7,5 +7,5 @@ class Donation(models.Model):
     donator = models.ForeignKey(Donator, on_delete=models.CASCADE)
     donation_type = models.CharField(max_length=10)
     donation_date = models.DateField()
-    donation_amount = models.IntegerField()
+    donation_amount = models.FloatField()
     donation_memo = models.CharField(max_length=255)
